[FEATURE] Disable uncaching in the Install Tool
The TYPO3 caching framework is not used by the Install Tool
except for the database compare function, which will create
tables according to the caching framework configuration.
It all caches are disabled the Install Tool wants to drop these
caching tables.

This fix disables the uncaching for Install Tool requests, therefor
caches are enabled and the cf_* tables won't get dropped.

resolves #13